refactor: Components 리팩토링
