+++ src/components/Button/Button.tsx
@@ -1,9 +1,5 @@
 import * as S from './style'
 
-export interface Props extends React.ButtonHTMLAttributes<HTMLButtonElement> {
-  color?: string
-}
-
-export function Button({ children, ...props }: Props) {
+export function Button({ children, ...props }: React.ButtonHTMLAttributes<HTMLButtonElement>) {
   return <S.Box {...props}>{children}</S.Box>
 }
