+++ src/components/Button/style.ts
@@ -1,16 +1,10 @@
-import styled, { css } from 'styled-components'
-import { MINT } from 'style/colors'
-import { Props } from './Button'
+import styled from 'styled-components'
+import { MINT } from 'styles/colors'
 
-export const Box = styled.button<Props>`
+export const Box = styled.button`
   border: 0;
   padding: 8px 12px;
   font-size: 16px;
   color: ${MINT};
   background: #fff;
-  ${({ color }) =>
-    color &&
-    css`
-      color: ${color};
-    `}
 `
