--- src/components/Card/BlankCard/index.tsx
@@ -1,1 +1,0 @@
-export { BlankCard } from './BlankCard'
