+++ src/components/Card/index.tsx
@@ -1,4 +1,3 @@
 import { Card } from './Card'
-import { BlankCard } from './BlankCard'
 import { DigitalCard } from './DigitalCard'
-export { Card, BlankCard, DigitalCard }
+export { Card, DigitalCard }
