+++ src/components/Form/FieldSet/CardSerialNumsFieldSet/CardSerialNumsFieldSet.tsx
@@ -14,7 +14,7 @@ export default function CardSerialNumsFieldSet({ serialNums, onChange }: Props) 
   return (
     <S.Box>
       <FieldSet legend="카드번호">
-        <S.Container>
+        <S.Wrapper>
           <S.Input type="number" name={first} value={firstVal} onChange={onChange} required />
           <S.Divider> - </S.Divider>
           <S.Input type="number" name={second} value={secondVal} onChange={onChange} required />
@@ -22,7 +22,7 @@ export default function CardSerialNumsFieldSet({ serialNums, onChange }: Props) 
           <S.Input type="password" name={third} value={thirdVal} onChange={onChange} required />
           <S.Divider> - </S.Divider>
           <S.Input type="password" name={fourth} value={fourthVal} onChange={onChange} required />
-        </S.Container>
+        </S.Wrapper>
       </FieldSet>
     </S.Box>
   )
