+++ src/hooks/useFieldSet.tsx
@@ -1,12 +1,5 @@
 import { useState } from 'react'
 
-export const useInput = (initialValue = '') => {
-  const [inputValue, setInput] = useState<string>(initialValue)
-  const handleInput = ({ target: { value } }: React.ChangeEvent<HTMLInputElement>) =>
-    setInput(value)
-  return [inputValue, handleInput] as const
-}
-
 export const useFieldSet = (initialValue: any) => {
   const [fieldsetValue, setFieldset] = useState(initialValue)
   const handleFieldSet = ({ target: { name, value } }: React.ChangeEvent<HTMLInputElement>) => {
